Replace stringly-typed enums with domain types and use NaiveDateTime for dates
- Use OrderChannel, OrderStatus, PaymentMethod enums instead of String
  in AdminOrder, SalesEvent, AdminReportSummary, and create_order()
- Add IouSettled variant to PaymentMethod for settled IOU orders
- Add Display, Copy, Hash, Ord impls and as_str() to domain enums
- Change Book.price_cents from u32 to i64 for consistency
- Replace String date fields with chrono::NaiveDateTime
- Rename occurred_on -> occurred_at, created_on -> created_at
- Rename current_utc_date() -> current_utc_datetime()

+++ crates/bookstore-cli/src/main.rs
@@ -24,7 +24,7 @@ enum Command {
         #[arg(long)]
         category: String,
         #[arg(long)]
-        price_cents: u32,
+        price_cents: i64,
     },
 }
